fix type double to big decimal on price of the food modal

+++ src/main/java/com/example/digitalmenu/dtos/FoodRequestDTO.java
@@ -1,4 +1,6 @@
 package com.example.digitalmenu.dtos;
 
-public record FoodRequestDTO(String title, String image, Double price) {
+import java.math.BigDecimal;
+
+public record FoodRequestDTO(String title, String image, BigDecimal price) {
 }
